requirements.txt
torch
torchvision
astropy
numpy
matplotlib
scikit-learn
umap-learn
hdbscan